+++ src/main/java/com/prgrms/tenwonmoa/config/WebSecurityConfig.java
@@ -10,7 +10,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.filter.CorsFilter;
 
-import com.prgrms.tenwonmoa.domain.user.jwt.JwtAuthenticationFilter;
+import com.prgrms.tenwonmoa.domain.user.security.jwt.JwtAuthenticationFilter;
 
 import lombok.RequiredArgsConstructor;
