+++ src/main/java/com/prgrms/tenwonmoa/domain/user/security/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.prgrms.tenwonmoa.domain.user.jwt;
+package com.prgrms.tenwonmoa.domain.user.security.jwt;
 
 import java.io.IOException;
