+++ src/main/java/com/prgrms/tenwonmoa/domain/user/service/UserService.java
@@ -10,7 +10,7 @@ import com.prgrms.tenwonmoa.domain.category.service.CreateDefaultUserCategorySer
 import com.prgrms.tenwonmoa.domain.user.User;
 import com.prgrms.tenwonmoa.domain.user.dto.CreateUserRequest;
 import com.prgrms.tenwonmoa.domain.user.dto.LoginUserResponse;
-import com.prgrms.tenwonmoa.domain.user.jwt.TokenProvider;
+import com.prgrms.tenwonmoa.domain.user.security.jwt.TokenProvider;
 import com.prgrms.tenwonmoa.domain.user.repository.UserRepository;
 import com.prgrms.tenwonmoa.exception.AlreadyExistException;
 import com.prgrms.tenwonmoa.exception.message.Message;
