+++ src/test/java/com/prgrms/tenwonmoa/domain/user/service/UserServiceTest.java
@@ -20,7 +20,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import com.prgrms.tenwonmoa.domain.category.service.CreateDefaultUserCategoryService;
 import com.prgrms.tenwonmoa.domain.user.User;
 import com.prgrms.tenwonmoa.domain.user.dto.CreateUserRequest;
-import com.prgrms.tenwonmoa.domain.user.jwt.TokenProvider;
+import com.prgrms.tenwonmoa.domain.user.security.jwt.TokenProvider;
 import com.prgrms.tenwonmoa.domain.user.repository.UserRepository;
 import com.prgrms.tenwonmoa.exception.AlreadyExistException;
 import com.prgrms.tenwonmoa.exception.message.Message;
